fix(bruno-request): refactor URL parsing to use parseUrl utility and remove legacy code
